requirements.txt
torch==2.4.1
torchvision==0.19.1
lightning
wandb 
hydra-core
hydra-colorlog
ipython
ipdb
rich
regex
tqdm
ftfy
scipy
scikit-metrics
autopep8
flake8
flake8-annotations
pylint
nltk